+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+* If it isn't possible to connect to a Unix Domain Socket a GvmError is raised
+  now [#207](https://github.com/greenbone/python-gvm/pull/207)
+
+### Fixed
+
+### Removed
+
+[Unreleased]: https://github.com/greenbone/python-gvm/compare/v1.4.0...master
+
 ## [1.4.0]
 
 ### Added