+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Added support for GMP 20.08 [#254](https://github.com/greenbone/python-gvm/pull/254)
+
+### Changed
+
+* Refactored Gmp classes into mixins [#254](https://github.com/greenbone/python-gvm/pull/254)
+
+[Unreleased]: https://github.com/greenbone/python-gvm/compare/v1.6.0...master
+
 ## [1.6.0] - 2020-06-10
 
 ### Added